feat(minifront): #1169: add Trade button to the assets table
